ImageEditorView  smashed
ImageEditorView has separated different subView
