+++ TestAppPhotoProcessing/LoadAndDisplay/DrawingCanvasView.swift
@@ -1,0 +1,20 @@
+//
+//  DrawingCanvasView.swift
+//  TestAppPhotoProcessing
+//
+//  Created by Андрей Завадский on 20.03.2025.
+//
+import SwiftUI
+import PencilKit
+
+struct DrawingCanvasView: UIViewRepresentable {
+    @Binding var canvasView: PKCanvasView
+    
+    func makeUIView(context: Context) -> PKCanvasView {
+        canvasView.drawingPolicy = .anyInput
+        return canvasView
+    }
+    
+    func updateUIView(_ uiView: PKCanvasView, context: Context) {}
+}
+
