+++ TestAppPhotoProcessing/LoadAndDisplay/ImageEditorView.swift
@@ -5,60 +5,7 @@
 //  Created by Андрей Завадский on 19.03.2025.
 //
 import SwiftUI
-import PhotosUI
 import PencilKit
-
-struct DrawingCanvasView: UIViewRepresentable {
-    @Binding var canvasView: PKCanvasView
-    
-    func makeUIView(context: Context) -> PKCanvasView {
-        canvasView.drawingPolicy = .anyInput
-        return canvasView
-    }
-    
-    func updateUIView(_ uiView: PKCanvasView, context: Context) {}
-}
-
-struct ImagePicker: UIViewControllerRepresentable {
-    @Binding var selectedImage: UIImage?
-    var sourceType: UIImagePickerController.SourceType = .photoLibrary
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator(self)
-    }
-    
-    func makeUIViewController(context: Context) -> UIImagePickerController {
-        let picker = UIImagePickerController()
-        picker.delegate = context.coordinator
-        picker.sourceType = sourceType
-        return picker
-    }
-    
-    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
-    
-    class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
-        var parent: ImagePicker
-        
-        init(_ parent: ImagePicker) {
-            self.parent = parent
-        }
-        
-        func imagePickerController(
-            _ picker: UIImagePickerController,
-            didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]
-        ) {
-            if let image = info[.originalImage] as? UIImage {
-                parent.selectedImage = image
-            }
-            picker.dismiss(animated: true)
-        }
-        
-        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
-            picker.dismiss(animated: true)
-        }
-    }
-}
-
 
 struct ImageEditorView: View {
     @State private var showImagePicker = false
@@ -95,8 +42,8 @@ struct ImageEditorView: View {
                         .rotationEffect(rotation)
                         .gesture(
                             MagnificationGesture()
-                                .onChanged { value in
-                                    scale = lastScale * value
+                                .onChanged {
+                                    scale = lastScale * $0
                                 }
                                 .onEnded { _ in
                                     lastScale = scale
@@ -104,8 +51,8 @@ struct ImageEditorView: View {
                         )
                         .gesture(
                             RotationGesture()
-                                .onChanged { angle in
-                                    rotation = lastRotation + angle
+                                .onChanged {
+                                    rotation = lastRotation + $0
                                 }
                                 .onEnded { _ in
                                     lastRotation = rotation
@@ -120,38 +67,15 @@ struct ImageEditorView: View {
                         .frame(height: 300)
                         .cornerRadius(16)
                         .shadow(radius: 8)
-                        .allowsHitTesting(true)
                 }
             }
             
-            VStack(spacing: 16) {
-                Button {
-                    sourceType = .photoLibrary
-                    showImagePicker = true
-                } label: {
-                    Label("Gallery", systemImage: "photo.on.rectangle")
-                }
-                
-                Button {
-                    sourceType = .camera
-                    showImagePicker = true
-                } label: {
-                    Label("Camera", systemImage: "camera")
-                }
-                
-                Button {
-                    showDrawing.toggle()
-                } label: {
-                    Label(showDrawing ? "Photo" : "Draw", systemImage: "pencil.tip")
-                }
-                
-                Button {
-                    saveEditedImage()
-                } label: {
-                    Label("Save", systemImage: "square.and.arrow.down")
-                }
-            }
-            .buttonStyle(.borderedProminent)
+            EditorControlsView(
+                sourceType: $sourceType,
+                showImagePicker: $showImagePicker,
+                showDrawing: $showDrawing,
+                saveAction: saveEditedImage
+            )
         }
         .sheet(isPresented: $showImagePicker) {
             ImagePicker(selectedImage: $selectedImage, sourceType: sourceType)
@@ -170,27 +94,24 @@ struct ImageEditorView: View {
     private func saveEditedImage() {
         guard let baseImage = selectedImage else { return }
         
-        // Объединяем рисунок и изображение
+        // объединяем рисунок и изображение
         let size = baseImage.size
         UIGraphicsBeginImageContextWithOptions(size, false, 0)
         baseImage.draw(in: CGRect(origin: .zero, size: size))
         
-        // Масштабируем canvas
+        // масштабируем canvas
         let drawing = canvasView.drawing.image(from: canvasView.bounds, scale: 1.0)
         drawing.draw(in: CGRect(origin: .zero, size: size))
         
         let finalImage = UIGraphicsGetImageFromCurrentImageContext()
         UIGraphicsEndImageContext()
         
-        guard let finalImage = finalImage else {
+        if finalImage != nil {
+            UIImageWriteToSavedPhotosAlbum(finalImage!, nil, nil, nil)
+            saveSuccess = true
+        } else {
             saveSuccess = false
-            showSaveAlert = true
-            return
         }
-        
-        // Сохраняем в фотопоток
-        UIImageWriteToSavedPhotosAlbum(finalImage, nil, nil, nil)
-        saveSuccess = true
         showSaveAlert = true
     }
 }
